+++ src/Exception/InvalidDataInputException.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace HiFolks\Statistics\Exception;
 
 use InvalidArgumentException;