+++ src/StudentT.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace HiFolks\Statistics;
 
 use HiFolks\Statistics\Exception\InvalidDataInputException;